new runtime: merge fatal.h into utils.h
