--- runtime/fatal.h
@@ -1,6 +1,0 @@
-#ifndef __FATAL_H_
-#define __FATAL_H_
-
-_Noreturn void fatal(const char* fmt, ...);
-
-#endif // __FATAL_H_
