+++ runtime/utils.h
@@ -2,5 +2,6 @@
 #define __UTILS_H_
 
 size_t minz(size_t a, size_t b);
+_Noreturn void fatal(const char* fmt, ...);
 
 #endif // __UTILS_H_
